requirements.txt
aiofiles==23.1.0
aiogram==3.0.0
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
annotated-types==0.5.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==4.1.0
bs4==0.0.1
bscscan-python==2.0.0
celery==5.3.4
celery-beatx==0.4.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.4.2
eventlet==0.33.3
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
kombu==5.3.2
loader==2017.9.11
magic-filter==1.0.11
motor==3.3.1
multidict==6.0.4
prompt-toolkit==3.0.39
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
rabbitmq==0.2.0
redis==5.0.0
requests==2.31.0
six==1.16.0
soupsieve==2.5
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
